Исправить реализацию обработчиков
