+++ src/server/adapters/LocalDateTimeAdapter.java
@@ -1,0 +1,33 @@
+package server.adapters;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonToken;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+    private final DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
+
+    @Override
+    public void write(JsonWriter writer, LocalDateTime value) throws IOException {
+        if (value == null) {
+            writer.nullValue();
+        } else {
+            writer.value(formatter.format(value.withNano(0)));
+        }
+    }
+
+    @Override
+    public LocalDateTime read(JsonReader reader) throws IOException {
+        if (reader.peek() == JsonToken.NULL) {
+            reader.nextNull();
+            return null;
+        }
+        return LocalDateTime.parse(reader.nextString(), formatter).withNano(0);
+    }
+}
+
