+++ src/server/handlers/TaskHandler.java
@@ -1,6 +1,5 @@
 package server.handlers;
 
-import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
 import com.sun.net.httpserver.HttpExchange;
 import tasks.Task;
@@ -10,11 +9,8 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 
 public class TaskHandler extends BaseHttpHandler {
-    private final Gson gson;
-
-    public TaskHandler(TaskManager taskManager, Gson gson) {
+    public TaskHandler(TaskManager taskManager) {
         super(taskManager);
-        this.gson = gson;
     }
 
     @Override
